// apps/web/src/api/generated/model/workbookSourceEditInvalidatesReferencesErrorResponseErrorDetails.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Lemma API
 * OpenAPI spec version: 0.1.0
 */
import type { WorkbookSourceEditInvalidatesReferencesErrorResponseErrorDetailsAffectedInsertedValuesItem } from "./workbookSourceEditInvalidatesReferencesErrorResponseErrorDetailsAffectedInsertedValuesItem.ts";

export type WorkbookSourceEditInvalidatesReferencesErrorResponseErrorDetails = {
  affectedInsertedValues: WorkbookSourceEditInvalidatesReferencesErrorResponseErrorDetailsAffectedInsertedValuesItem[];
  recoveryAction: string;
  summary: string;
};